[Ref] Remove c99_snprintf and c99_vsnprintf which confusingly were not 100% c99 compliant. Replace the one single usage site and always zero-terminate strings by hand there. [Fix] Fix 64bit warning in libopenmpt_modplug. [Ref] Silence stupid 64bit warning in ITTools.cpp.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@2149 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/ITTools.cpp
@@ -177,7 +177,7 @@ void ITOldInstrument::ConvertToMPT(ModInstrument &mptIns) const
 	mptIns.VolEnv.nNodes = 25;
 
 	// Volume Envelope Data
-	for(size_t i = 0; i < 25; i++)
+	for(uint32 i = 0; i < 25; i++)
 	{
 		if((mptIns.VolEnv.Ticks[i] = nodes[i * 2]) == 0xFF)
 		{
